DB Change for Disable Vehicle - Jasmeet

+++ DealWheel/src/servlet/DeleteVehicle.java
@@ -5,6 +5,8 @@ import static services.utility.GenericConstant.NAV_TO_VENDOR_HOME_PAGE;
 
 import java.io.IOException;
 import java.math.BigInteger;
+import java.util.ArrayList;
+import java.util.List;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -12,9 +14,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import model.Vehicle;
+
 import org.apache.log4j.Logger;
 
-import model.Vehicle;
 import dao.VehicleDAOImpl;
 
 @WebServlet("/DeleteVehicle")
@@ -29,9 +32,22 @@ public class DeleteVehicle extends HttpServlet {
 		try{
 			VehicleDAOImpl<Vehicle> vDAO = new VehicleDAOImpl<Vehicle>();
 			Object[] list=  req.getParameterValues("arrayList");
-			for (int i = 0; i < list.length; i++) {
-			if("Yes".equals(req.getParameter("check" + i)))
-				vDAO.delete(new BigInteger(list[i].toString()));
+			if("Delete".equals(req.getParameter("identifier"))){
+				for (int i = 0; i < list.length; i++) {
+					if("Yes".equals(req.getParameter("check" + i)))
+						vDAO.delete(new BigInteger(list[i].toString()));
+					}
+			}
+			else if("Disable".equals(req.getParameter("identifier"))){
+				List<BigInteger> listIds = new ArrayList<BigInteger>();;
+				for (int i = 0; i < list.length; i++) {
+					if("Yes".equals(req.getParameter("check" + i)))
+						listIds.add(new BigInteger(list[i].toString()));
+					}
+				vDAO.disable(listIds);
+			}
+			else {
+				vDAO.enable(new BigInteger(req.getParameter("identifier")));
 			}
 			resp.sendRedirect(req.getContextPath()+NAV_TO_DISPLAY_VEHICLE_PAGE);
 		}catch(Exception e){
